chore: refactor benchmarks structure
