+++ .github/workflows/benchmark-comparison.yml
@@ -42,13 +42,13 @@ jobs:
           develop --impure --command just
           --justfile ./test/performance/justfile
           --working-directory ./test/performance 
-          compare ${{ inputs.bench }} ${{ inputs.parallelism }} ${{ inputs.duration }} ${{ inputs.count }}
+          writes compare ${{ inputs.bench }} ${{ inputs.parallelism }} ${{ inputs.duration }} ${{ inputs.count }}
       - run: >
           /nix/var/nix/profiles/default/bin/nix --extra-experimental-features "nix-command" --extra-experimental-features "flakes"
           develop --impure --command just 
           --justfile ./test/performance/justfile 
           --working-directory ./test/performance 
-          graphs
+          writes graphs
       - uses: actions/upload-artifact@v4
         with:
           name: graphs
