+++ test/performance/justfile
@@ -1,42 +1,47 @@
-set dotenv-load
-set positional-arguments
+#import 'pkg/write/justfile'
 
-tmpdir := `mktemp -d`
+mod writes 'pkg/write/justfile'
 
-# Run against the local branch
-run bench='.' p='1' benchtime='1s' count='1' output='./report/benchmark-output.txt':
-    mkdir -p $(dirname {{output}})
-    rm -f {{output}}
-    go test -run ^$ -tags it,local \
-        -report.file ./report/report.json \
-        -timeout 600m \
-        -bench={{bench}} \
-        -count={{count}} \
-        -parallelism {{p}} \
-        -test.benchtime {{benchtime}} . | tee -a {{output}}
 
-run-remote ledgerURL bench='.' p='1' benchtime='1s' count='1' output='./report/benchmark-output.txt':
-    mkdir -p $(dirname {{output}})
-    rm -f {{output}}
-    go test -run ^$ -tags it,local \
-        -report.file ./report/report.json \
-        -ledger.url {{ledgerURL}} \
-        -timeout 600m \
-        -bench={{bench}} \
-        -count={{count}} \
-        -parallelism {{p}} \
-        -test.benchtime {{benchtime}} . | tee -a {{output}}
-
-# Compare against main branch
-compare bench='.' p='1' benchtime='1s' count='1' output='./report/benchmark-output.txt':
-    trap 'rm -rf {{tmpdir}}' EXIT
-    just run {{bench}} {{p}} {{benchtime}} {{count}} './report/benchmark-output-local.txt'
-    rm -f ./report/benchmark-output-main.txt
-    git clone --depth 1 -b main https://github.com/formancehq/ledger {{tmpdir}}
-    location=$(pwd) && cd {{tmpdir}}/test/performance && just run {{bench}} {{p}} {{benchtime}} {{count}} $location/report/benchmark-output-main.txt
-    benchstat ./report/benchmark-output-main.txt ./report/benchmark-output-local.txt > ./report/benchmark-comparison.txt || true
-
-graphs:
-    cd charts && npm install
-    cd charts && npm run build
-    cd charts && node ./index.js
+#set dotenv-load
+#set positional-arguments
+#
+#tmpdir := `mktemp -d`
+#
+## Run against the local branch
+#writes bench='.' p='1' benchtime='1s' count='1' output='./report/writes/benchmark-output.txt':
+#    mkdir -p $(dirname {{output}})
+#    rm -f {{output}}
+#    go test -run ^$ -tags it,local \
+#        -timeout 600m \
+#        -bench={{bench}} \
+#        -count={{count}} \
+#        -test.benchtime {{benchtime}} ./pkg/write/... \
+#            -report.file $(pwd)/report/writes/report.json \
+#            -parallelism {{p}} | tee -a {{output}}
+#
+#writes-remote ledgerURL bench='.' p='1' benchtime='1s' count='1' output='./report/writes/benchmark-output.txt':
+#    mkdir -p $(dirname {{output}})
+#    rm -f {{output}}
+#    go test -run ^$ -tags it,local \
+#        -timeout 600m \
+#        -bench={{bench}} \
+#        -count={{count}} \
+#        -test.benchtime {{benchtime}} ./pkg/write/... \
+#            -ledger.url {{ledgerURL}} \
+#            -report.file $(pwd)/report/writes/report.json \
+#            -parallelism {{p}} | tee -a {{output}}
+#
+## Compare against main branch
+#writes-compare bench='.' p='1' benchtime='1s' count='1' output='./report/writes/benchmark-output.txt':
+#    trap 'rm -rf {{tmpdir}}' EXIT
+#    just run {{bench}} {{p}} {{benchtime}} {{count}} './report/writes/benchmark-output-local.txt'
+#    rm -f ./report/writes/benchmark-output-main.txt
+#    git clone --depth 1 -b main https://github.com/formancehq/ledger {{tmpdir}}
+#    location=$(pwd) && cd {{tmpdir}}/test/performance && just run {{bench}} {{p}} {{benchtime}} {{count}} $location/report/writes/benchmark-output-main.txt
+#    benchstat ./report/writes/benchmark-output-main.txt ./report/writes/benchmark-output-local.txt > ./report/writes/benchmark-comparison.txt || true
+#
+#writes-graphs:
+#    cd charts && npm install
+#    cd charts && npm run build
+#    cd charts && node ./index.js
