+++ test/performance/pkg/write/charts/src/colors.ts
