+++ test/performance/pkg/write/charts/tsconfig.json
