+++ test/performance/pkg/write/example_scripts/example1.js
