+++ test/performance/pkg/write/report_test.go
@@ -1,10 +1,10 @@
 //go:build it
 
-package performance_test
+package write_test
 
 import (
 	"github.com/formancehq/go-libs/v2/time"
-	"github.com/formancehq/ledger/test/performance/env"
+	"github.com/formancehq/ledger/test/performance/pkg/env"
 	"github.com/jamiealquiza/tachymeter"
 	"sync"
 )
