+++ test/performance/pkg/write/scripts/any_to_bank.js
