+++ test/performance/pkg/write/scripts/world_to_bank.js
